+++ app/src/main/java/mihon/core/migration/migrations/Migrations.kt
@@ -5,7 +5,7 @@ import mihon.core.migration.Migration
 val migrations: List<Migration>
     get() = listOf(
         SetupBackupCreateMigration(),
-        SetupLibraryUpdateMigration(),
+        LegacyLibraryUpdateCleanupMigration(),
         TrustExtensionRepositoryMigration(),
         CategoryPreferencesCleanupMigration(),
         InstallationIdMigration(),